update game
add game.py
move code from main.py to game.py
create new Game class in game.py
add background sound and update new shoot_fx and explosion_fx
